Move handleWheel to custom hook
